requirements.txt
asgiref==3.2.10
Django==3.1
django-background-tasks==1.2.5
django-compat==1.0.15
django-crispy-forms==1.9.2
django-widget-tweaks==1.4.8
psycopg2==2.8.5 --no-binary=psycopg2
pytz==2020.1
sh==1.13.1
six==1.15.0
sqlparse==0.3.1
django-cryptography